requirements.txt
requests>=2.31.0
xmltodict>=0.13.0
google-generativeai>=0.7.2
